requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
alembic==1.16.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
azure-core==1.32.0
azure-storage-blob==12.24.0
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==43.0.3
cssselect2==0.8.0
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.6
fastjsonschema==2.21.1
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google_search_results==2.4.2
greenlet==3.2.2
h11==0.14.0
html2text==2025.4.15
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-pinecone==0.2.5
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==5.3.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
multidict==6.4.3
munch==2.5.0
mypy_extensions==1.1.0
mysql-connector-python==9.2.0
narwhals==1.44.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo4j==5.28.1
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.2.5
openai==1.77.0
openevals==0.1.0
opentelemetry-api==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20250506
pexpect==4.9.0
pillow==11.0.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
playwright==1.52.0
plotly==5.24.1
pluggy==1.5.0
postmarker==1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.5
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.27.1
pydeck==0.9.1
PyDocX==0.9.10
pydyf==0.11.0
pyee==13.0.0
Pygments==2.18.0
PyJWT==2.10.1
motor==3.7.1
pymongo==4.10.1
pypdf==5.4.0
pyphen==0.17.2
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==3.3.0
python-multipart==0.0.19
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.25.1
rsa==4.9
scikit-learn==1.7.0
scipy==1.16.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.37.2
streamlit==1.46.1
syrupy==4.9.1
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trulens-apps-langchain==1.5.2
trulens-core==1.5.2
trulens-dashboard==1.5.2
trulens-feedback==1.5.2
trulens-otel-semconv==1.5.2
trulens-providers-openai==1.5.2
trulens_eval==1.5.2
typer==0.15.1
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
wcwidth==0.2.13
weasyprint==66.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
widgetsnbextension==4.0.14
wrapt==1.17.2
XlsxWriter==3.2.0
yarl==1.20.0
zenrows==1.4.0
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.23.0
